refactor(goal): move vue-only helpers out of application layer
- Switch StatusRuleEditor to import sortRulesByPriority from @dailyuse/goal/application-client
- Switch TemplateBrowser to import GoalTemplate type from @dailyuse/goal/application-client
- TemplateRecommendationService now consumes BUILT_IN_TEMPLATES from package instead of empty local array
- Move DAGExportService, GoalTimelineService, TemplateRecommendationService, WeightRecommendationService to utils/
- Move useAutoStatusRules, useWeightSnapshot composables from application/composables/ to composables/
- Update all Vue component imports to new locations
- Delete application/ directory (no longer needed)
- Add utils/index.ts barrel export
- Update composables/index.ts barrel to include new composables

Verified: app-vue:typecheck, app-vue:test (46 files, 136 tests), goal:test (54 files, 297 tests),
goal:typecheck, vue-tsc web, vue-tsc desktop, governance-check
