--- packages/app-vue/src/modules/goal/application/rules/BuiltInRules.ts
@@ -1,5 +1,0 @@
-import type { StatusRule } from '@dailyuse/contracts/goal';
-
-export const sortRulesByPriority = (rules: StatusRule[]) => {
-  return [...rules].sort((left, right) => right.priority - left.priority);
-};
