+++ packages/app-vue/src/modules/goal/components/dag/ExportDialog.vue
@@ -94,7 +94,7 @@
 <script setup lang="ts">
 import { ref, computed, type Component } from 'vue';
 import { useI18n } from 'vue-i18n';
-import type { ExportOptions } from '../../application/services/DAGExportService';
+import type { ExportOptions } from '../../utils/dag-export';
 import {
   Dialog,
   DialogContent,
