+++ packages/app-vue/src/modules/goal/components/timeline/GoalTimelineView.vue
@@ -102,7 +102,7 @@ const { t } = useI18n();
 import type { ECharts } from 'echarts';
 import TimelineControls from './TimelineControls.vue';
 import { useGoalTimeline } from '../../composables/useGoalTimeline';
-import type { TimelineSnapshot } from '../../application/services/GoalTimelineService';
+import type { TimelineSnapshot } from '../../composables/types';
 
 // ==================== Props ====================
