+++ packages/app-vue/src/modules/goal/components/timeline/TimelineControls.vue
@@ -138,8 +138,8 @@
 <script setup lang="ts">
 import { ref, computed, watch, onUnmounted } from 'vue';
 import { useI18n } from 'vue-i18n';
-import type { TimelineSnapshot } from '../../application/services/GoalTimelineService';
-import { formatTimelineTimestamp } from '../../application/services/GoalTimelineService';
+import type { TimelineSnapshot } from '../../composables/types';
+import { formatTimelineTimestamp } from '../../composables/types';
 
 const { t } = useI18n();
