+++ packages/app-vue/src/modules/goal/composables/index.ts
@@ -4,5 +4,7 @@
 
 export { useGoal } from './useGoal';
 export { useGoalTimeline } from './useGoalTimeline';
+export { useAutoStatusRules } from './useAutoStatusRules';
+export { useWeightSnapshot } from './useWeightSnapshot';
 export type { TimelineData, TimelineSnapshot } from './types';
 export { formatTimelineTimestamp } from './types';
