+++ packages/app-vue/src/modules/goal/composables/useAutoStatusRules.ts
