+++ packages/app-vue/src/modules/goal/composables/useWeightSnapshot.ts
