+++ packages/app-vue/src/modules/goal/utils/index.ts
@@ -1,0 +1,13 @@
+export { dagExportService } from './dag-export';
+export type { ExportFormat, ExportOptions } from './dag-export';
+
+export type { TimelineSnapshot, TimelineData } from './goal-timeline';
+export { formatTimelineTimestamp } from './goal-timeline';
+
+export { default as templateRecommendationService } from './template-recommendation';
+export type { RecommendationFilters } from './template-recommendation';
+
+export { weightRecommendationService } from './weight-recommendation';
+export type { WeightStrategy } from './weight-recommendation';
+
+export * from './progress';
